// convex/_generated/api.d.ts
declare const fullApi: ApiFromModules<{
  admin: typeof admin;
  analytics: typeof analytics;
  debug: typeof debug;
  fixSubscription: typeof fixSubscription;
  http: typeof http;
  permissions: typeof permissions;
  polarApi: typeof polarApi;
  polarSync: typeof polarSync;
  products: typeof products;
  retell: typeof retell;
  retellConfig: typeof retellConfig;
  security: typeof security;
  settings: typeof settings;
  setup: typeof setup;
  subscriptions: typeof subscriptions;
  userSync: typeof userSync;
  users: typeof users;
}>;
declare const fullApiWithMounts: typeof fullApi;

export declare const api: FilterApi<
  typeof fullApiWithMounts,
  FunctionReference<any, "public">
>;